requirements.txt
# Core ML and Data Science
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# NLP for report parsing
nltk>=3.8.0

# Document parsing
python-docx>=1.1.0
python-multipart>=0.0.6

# Geospatial analysis
geopy>=2.4.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
jinja2>=3.1.0

# HTTP requests for data collection
requests>=2.31.0
beautifulsoup4>=4.12.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Date/time handling
python-dateutil>=2.8.0

# Scheduling
schedule>=1.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Claude AI integration
anthropic>=0.39.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
